cpp File

+++ cpp File/PG67257/p67257.cpp
@@ -61,7 +61,7 @@ void solve(int depth)
         if (visited[i]) continue;
         
         visited[i] = true;
-        opvec.push_back((ll)i);
+        opvec.push_back(i);
         
         solve(depth + 1);
